Individual view to drf (#477)
* ACT-725 When Editing a project other organizations' programs are available

Current Behavior: All the programs in the system are available when editing a project

Expected Behavior: Only programs belonging to an organization should be available when editing a projects

* ACT-727 Change Django views to DRF views: Individual model

Left the previous views for now. If this is ok, I can proceed to integrate with the front-end.
Also, having one view for both List+CRUD is no good practice. so, my changes can be tested at
/formlibrary/individual for List
and
formlibrary/individual/PK for single items

Co-authored-by: Andrew <andrew.tc.pham@gmail.com>